refactor(StoriesControllerMetrics): module removed
This module has been dismatled as it was written for stories#show
Method and after action hook moved to StoriesController
