+++ app/controllers/supplejack_api/concerns/stories_controller_metrics.rb
@@ -1,25 +1,25 @@
 # frozen_string_literal: true
 
-module SupplejackApi
-  module Concerns
-    module StoriesControllerMetrics
-      extend ActiveSupport::Concern
-      include IgnoreMetrics
+# module SupplejackApi
+#   module Concerns
+#     module StoriesControllerMetrics
+#       extend ActiveSupport::Concern
+#       include IgnoreMetrics
 
-      included do
-        after_action :create_story_record_views, only: :show
+#       included do
+#         after_action :create_story_record_views, only: :show
 
-        def create_story_record_views
-          return unless log_request_for_metrics?
+#         def create_story_record_views
+#           return unless @api_response[:payload] && log_request_for_metrics?
 
-          payload = JSON.parse(response.body)
-          log = payload['contents']&.map do |record|
-            { record_id: record['record_id'], display_collection: record['content']['display_collection'] }
-          end
-
-          SupplejackApi::RequestMetric.spawn(log, 'user_story_views') if log
-        end
-      end
-    end
-  end
-end
+#           SupplejackApi::RequestMetric.spawn(
+#             @api_response[:payload][:contents].map do |record|
+#               { record_id: record[:record_id], display_collection: record[:content][:display_collection] }
+#             end,
+#             'user_story_views'
+#           )
+#         end
+#       end
+#     end
+#   end
+# end
