+++ spec/controllers/supplejack_api/concerns/stories_controller_metrics_spec.rb
@@ -1,25 +1,25 @@
-require 'spec_helper'
+# require 'spec_helper'
 
-describe ApplicationController, type: :controller do
-  controller do
-    include SupplejackApi::Concerns::StoriesControllerMetrics
+# describe ApplicationController, type: :controller do
+#   controller do
+#     include SupplejackApi::Concerns::StoriesControllerMetrics
 
-    def show
-      @api_response = { payload:  ::StoriesApi::V3::Presenters::Story.new.call(SupplejackApi::UserSet.first) }
-      head :ok
-    end
-  end
+#     def show
+#       @api_response = { payload:  ::StoriesApi::V3::Presenters::Story.new.call(SupplejackApi::UserSet.first) }
+#       head :ok
+#     end
+#   end
 
-  let!(:story) { create(:story) }
+#   let!(:story) { create(:story) }
 
-  describe 'GET#show' do
-    it 'creates a user_story_views SupplejackApi::RequestMetric' do
-      get :show, params: { id: 1 }
+#   describe 'GET#show' do
+#     it 'creates a user_story_views SupplejackApi::RequestMetric' do
+#       get :show, params: { id: 1 }
 
-      expect(SupplejackApi::RequestMetric.count).to eq 1
-      expect(SupplejackApi::RequestMetric.first.records.map { |x| x[:record_id] }).to eq story.set_items.map(&:record_id)
-      expect(SupplejackApi::RequestMetric.first.records.map { |x| x[:display_collection] }).to eq story.set_items.map { |x| x[:content][:display_collection] }
-      expect(SupplejackApi::RequestMetric.first.metric).to eq 'user_story_views'
-    end
-  end
-end
+#       expect(SupplejackApi::RequestMetric.count).to eq 1
+#       expect(SupplejackApi::RequestMetric.first.records.map { |x| x[:record_id] }).to eq story.set_items.map(&:record_id)
+#       expect(SupplejackApi::RequestMetric.first.records.map { |x| x[:display_collection] }).to eq story.set_items.map { |x| x[:content][:display_collection] }
+#       expect(SupplejackApi::RequestMetric.first.metric).to eq 'user_story_views'
+#     end
+#   end
+# end
